+++ tests/JWT.Tests.Common/Algorithms/ECDSAAlgorithmTests.cs
@@ -3,7 +3,6 @@ using System;
 using System.Collections.Generic;
 using System.Security.Cryptography;
 using System.Security.Cryptography.X509Certificates;
-using System.Text;
 using FluentAssertions;
 using JWT.Algorithms;
 using JWT.Tests.Models;
@@ -57,8 +56,7 @@ namespace JWT.Tests.Algorithms
         [DataTestMethod]
         public void Ctor_Should_Not_Throw_Exception_When_Certificate_Has_No_PrivateKey(Func<X509Certificate2, ECDSAAlgorithm> algFactory)
         {
-            var bytes = Convert.FromBase64String(TestData.ServerEcdsaPublicKey);
-            var cert = new X509Certificate2(bytes);
+            var cert = TestData.CertificateWithPublicKeyEcdsa;
 
             var alg = algFactory(cert);
